Configured the user page and added login and register buttons
I imported some basic website template files into my project and split up the code into separate blade.php files. I also connected all the routes to the new user page login and register buttons, including an if condition t display the login and register buttons when a user is logged out, and a logout button when the user is logged in.