Added an orders page for the admin user to allow them to process online orders.
In the admin panel I created a new page for displaying orders made by normal users using the sight. The table allows the admin to view all orders made and process them by clicking a button next to the item in the table. Once the button is clicked the order goes from a processing status to a delivered status.